separate ApplicationManager.cs from LightingStateManager.cs
