+++ Project-Aurora/Project-Aurora/Controls/GameStateParameterPicker.xaml.cs
@@ -9,6 +9,7 @@ using System.Windows.Data;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Animation;
+using AuroraRgb.EffectsEngine;
 using AuroraRgb.Profiles;
 using AuroraRgb.Utils;
 using Xceed.Wpf.Toolkit;
