+++ Project-Aurora/Project-Aurora/Profiles/Dota 2/Dota2Profile.cs
@@ -1,4 +1,5 @@
 ﻿using System.Drawing;
+using AuroraRgb.EffectsEngine;
 using AuroraRgb.Profiles.Dota_2.Layers;
 using AuroraRgb.Settings;
 using AuroraRgb.Settings.Layers;
