+++ Project-Aurora/Project-Aurora/Settings/Overrides/Logic/Boolean/Boolean_GameState.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Globalization;
 using System.Windows.Media;
+using AuroraRgb.EffectsEngine;
 using AuroraRgb.Profiles;
 using AuroraRgb.Utils.Json;
 using Newtonsoft.Json;
